requirements.txt
django
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers
uvicorn